requirements/requirements.txt
requests
pandas
psycopg2
psycopg2-binary
apache-airflow 
